Added basic travis CI.
This needs a lot of work with unit tests actually writing etc before this becomes useful. But this is a hack day and being done for a demo at a local event I am hosting. I'll write some proper unit tests once I have some more time on this again.